Update pgjwt documentation with link to Supabase JWT Handling Documentation
Added notes about JWT handling in Supabase. It might seem stupid but for whatever reason I though I needed the extension turned on and delayed upgrading my database because of this. I read through the documentation and finally realized it has zero impact on the JWT functionality built in. Figured someone else out there may not realize this and is delaying upgrading theis postgres version, so I added a note at the top clarifying how Supabase handles JWT with a link to the documentation.