refactor(close confirmation modal): use new useConfirmOnClose hook
We have a repeated pattern many places, where we have a modal or side
panel that should show a confirmation modal before closing -- but only
if the user has made changes. This is repeated everywhere it's needed,
with the logic mixed in alongside everything else in the panel/modal.

This refactors all the instances to use the useConfirmOnClose hook
instead, which should make it more encapsulated and explicit what is
happening.