fix(cron): optimize query for cron tables
The `cron.job_run_details` table can get very long (millions of rows).
It's only indexed on the `runid` column, and the Postgres team considers
it a difficult task to add new indexes.

On the job run details table, we automatically sort and paginate by the
`start_time` column, so this creates a problem for very large tables.
This PR changes to use the `runid` column instead. Since `runid`s are
sequential, this should correlate pretty closely with `start_time`
anyway, while letting Postgres use the index.

Similarly, on the cron overviews table, the search for the latest run
required a sequential scan on the `job_run_details` table. Changed this
to also use the `runid` index instead.