fix(table editor): unnecessary column updates (#40446)
* fix(table editor): unnecessary column updates

When you use the table editor to edit a table containing array columns,
even if you _make no changes_, an update is sent to alter the array
column's type to its original type. This happens because we aren't
correctly checking for array column equality when we diff old/new column
configs. Fixed this so no update will be sent if the column isn't
changed.

* Nit refactor

* fix(table editor): bug with editing column twice in a row

---------

Co-authored-by: Joshen Lim <joshenlimek@gmail.com>