feat(docs): content api allow get requests for caching (#41443)
Content API requests were all accepted via POST (as detailed in the
GraphQL spec) and not amenable to caching.

Now, requests are also accepted via GET (as optionally allowed for query
requests only), and cached.

Also sent the request method to Logflare so we can track the ratio of
GET/POST requests.